fix: resolve cross-file import for _make_tool_use_line, fix ruff lint in arch guard
Move _make_tool_use_line from test_headless_path_validation to test_headless_core
so synthesis can import it from a canonical location. Fix E741 ambiguous variable
name in test_headless_split.py arch guard.
